компонент Logout, UserData
