+++ src/components/Logout/Logout.jsx
@@ -1,0 +1,12 @@
+import { LogOut, LogOutIcon } from "./Logout.styled";
+
+const Logout = () => {
+  return (
+    <LogOut>
+      <LogOutIcon />
+      log out
+    </LogOut>
+  );
+};
+
+export default Logout;
