+++ src/components/Logout/Logout.styled.js
@@ -1,0 +1,38 @@
+import styled from "styled-components";
+import { HiArrowRightOnRectangle } from "react-icons/hi2";
+
+export const LogOut = styled.button`
+  position: absolute;
+  right: 12px;
+  bottom: 20px;
+  display: inline-flex;
+  align-items: center;
+  margin-left: auto;
+  margin-bottom: 8px;
+  border: none;
+  background-color: transparent;
+  color: ${p => p.theme.colors.greyText};
+  font-size: 16px;
+  font-weight: 500;
+  cursor: pointer;
+  :hover svg,
+  :focus svg {
+    fill: ${p => p.theme.colors.hoverBtn};
+  }
+
+  @media ${p => p.theme.media.tablet} {
+    left: 32px;
+    bottom: 24px;
+    margin: 0;
+    padding: 0;
+  }
+  @media ${p => p.theme.media.desktop} {
+    left: 16px;
+  }
+`;
+export const LogOutIcon = styled(HiArrowRightOnRectangle)`
+  width: 18px;
+  height: 18px;
+  margin-right: 8px;
+  color: ${p => p.theme.colors.accent};
+`;
