+++ src/components/UserData/UserData.jsx
@@ -1,10 +1,10 @@
 import AddPetBtn from "components/AddPetBtn";
 import UserDataItem from "components/UserDataItem";
-import { BoxUser, TitleUser, TitleBtn, BoxTitleUser, BoxBtnTablet } from "./UserData.styled";
+import { TitleUser, TitleBtn, BoxTitleUser, BoxBtnTablet } from "./UserData.styled";
 
 const UserData = () => {
   return (
-    <BoxUser>
+    <>
       <BoxTitleUser>
         <TitleUser>My information:</TitleUser>
         <BoxBtnTablet>
@@ -13,7 +13,7 @@ const UserData = () => {
         </BoxBtnTablet>
       </BoxTitleUser>
       <UserDataItem />
-    </BoxUser>
+    </>
   );
 };
