+++ src/pages/UserPage/UserPage.jsx
@@ -1,14 +1,18 @@
+import Logout from "components/Logout";
 import PetsList from "components/PetsList";
 import UserData from "components/UserData";
 
-import { UserPageContainer } from "./UserPage.styled";
+import { UserPageContainer, BoxUser } from "./UserPage.styled";
 
 const UserPage = () => {
   return (
     <>
       <div style={{ height: "42px", backgroundColor: "green", marginBottom: "61px" }}></div>
       <UserPageContainer>
-        <UserData />
+        <BoxUser>
+          <UserData />
+          <Logout />
+        </BoxUser>
         <PetsList />
       </UserPageContainer>
     </>
